Add SPI interface and token CLI implementation
Defined a simple first-pass SPI for invoking code clone detection
on various strategies, which can be linked in from separate jars,
compiled using the SPI jar.

Created a simple service provider, TestCC, which will respond with
random 'discoveries' after a random delay. This should be useful
for testing the respoinsiveness of developed UIs.

Soon the SourcererCCJ project will be compiled in the same way,
and will become a service provider for the CLI and UI elements
of this repo