Integrate sourcerercc strategy, add search status
Changed the SPI definition in order to notify the main app when a
strategy is done working on a search.

Integrted the sourcerercc strategy provider, it works quite
nicely (and apparently very quickly on the toy test cases
I've worked on). This may need to be updated again when
I finish adding language support.